triangulo

+++ TRIANGULO.py
@@ -1,5 +1,5 @@
-base=float(input("ingresa base:"))
-altura=float(input("ingrese altura:"))
+base=int(input("ingresa base:"))
+altura=int(input("ingrese altura:"))
 area=(base * altura)/2
 print("area es:",area)
